added friends page to only allow you to add other people as friends and not duplicates
